Sửa các thuộc tính của Packet

+++ src/protocol/Packet.java
@@ -6,10 +6,10 @@ public class Packet implements Serializable {
     // Cau truc packet don gian
 
     private Message msgType;
-    private long dataLength;
+    private int dataLength;
     private byte[] data = new byte[1024];
 
-    public Packet(Message msgType, long dataLength, byte[] data) {
+    public Packet(Message msgType, int dataLength, byte[] data) {
         this.msgType = msgType;
         this.dataLength = dataLength;
         this.data = data;
@@ -27,11 +27,11 @@ public class Packet implements Serializable {
         this.msgType = msgType;
     }
 
-    public long getDataLength() {
+    public int getDataLength() {
         return dataLength;
     }
 
-    public void setDataLength(long dataLength) {
+    public void setDataLength(int dataLength) {
         this.dataLength = dataLength;
     }
